pyproject.toml
[project]
name = "mrna-plum"
version = "0.1.0"
description = "mRNA-PLUM CLI (merge/parse/stats/export) for Moodle/PLUM logs"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0",
  "duckdb>=1.0.0",
  "pandas>=2.0.0",
  "pyarrow>=15.0.0",
  "openpyxl>=3.1.0",
]

[project.scripts]
mrna_plum = "mrna_plum.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
